requirements.txt
pandas==2.2.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
langchain-community==0.2.0
openai==1.47.1
numpy==1.26.4
scikit-learn==1.5.2
tqdm==4.66.5
torch==2.2.2
transformers==4.44.2
matplotlib==3.9.2